+++ docs/changelog.rst
@@ -28,6 +28,10 @@ Bug fix release.
   and docs to match how the core use the backend's browse method. (Fixes:
   :issue:`833`)
 
+- Local library API: Add :attr:`mopidy.local.ROOT_DIRECTORY_URI` constant for
+  use by implementors of :method:`mopidy.local.Library.browse`. (Related to:
+  :issue:`833`)
+
 
 v0.19.3 (2014-08-03)
 ====================